+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Next Version
 ## Misc
 - Switch from jsx to [erl-json](https://github.com/galdor/erl-json).
+- For JSON logs, format the domain as a single string where domain parts are
+  separated by full stop characters.
 
 # 1.1.0
 ## Bugs